+++ attributes/default.rb
@@ -1,5 +1,3 @@
 id = 'redirect'
 
-default[id]['hsts_max_age'] = 15_768_000
-
 default[id]['hosts'] = []